rename outlier.js to outlierTransformer.js
