+++ lib/statistics/index.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const idwTransformer = require('./idw/idwTransformer'),
-  outlier = require('./outlier/outlier');
+  outlierTransformer = require('./outlier/outlierTransformer');
 
 module.exports = {
   idwTransformer,
-  outlier
+  outlierTransformer
 };
