+++ lib/statistics/outlier/outlierTransformer.js
